obs/lightcone{hod} -> lightcone/hod{hod:03}
