+++ cmass/summary/tools.py
@@ -48,10 +48,10 @@ def get_nofz(z, fsky, cosmo=None):
 @timing_decorator
 def load_lightcone(source_dir, hod_seed, filter_name=None):
     if filter_name is None:
-        infile = pjoin(source_dir, 'obs', f'lightcone{hod_seed}.h5')
+        infile = pjoin(source_dir, 'lightcone', f'hod{hod_seed:03}.h5')
     else:
         infile = pjoin(source_dir, 'filter',
-                       f'lightcone{hod_seed}_{filter_name}.h5')
+                       f'hod{hod_seed:03}_{filter_name}.h5')
 
     with h5py.File(infile, 'r') as f:
         ra = f['ra'][...]
